+++ docker/aggregator-import-servers.sh
@@ -2,5 +2,5 @@
 set -o xtrace
 set -e
 
-/bin/bash /tmp/write-config.sh
+/usr/bin/php /var/www/html/main_server/artisan settings:sync --force
 /usr/bin/php /var/www/html/main_server/artisan aggregator:ImportRootServers